initializers: moved some initializers into the background job
The `portus_user.rb` and `registry.rb` initializers have been moved into a new
background job named `Initializer`. This has been done so Rails 6.x
doesn't (rightfully) complain that we are using app constants inside of the
`config` directory (and thus messing with the autoload procedure).

This new background task is only called at the very beginning (thus, before
running the actual execution loop). This way we achieve the same effect as if
they were inside of the `config` directory.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>