config: suppress warning on multiple databases
As of Rails 6.x you can setup multiple databases to be used on the same
setup. That being said, Rails was bailing on detecting that in our case because
we abuse the ERB idea on the `config/database.yml` file.

This commit simplifies that and jots down that we should check whether mutltiple
databases is something that might be configurable.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>