Update commit links in changelog
Github no longer detects git commits in markdown documents and transforms them
into links. Thus, the SHAs are not that useful from a UX point of view on the
changelog file. I've updated all these SHAs so they are also contain a link to
the Github page containing the commit.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>