Moved to webpacker
This involved a lot of heavy-lifting, since we had to rethink the whole webpack
setup. That being said, this is needed as of Rails 6, since webpacker is not the
official way to manage JS code and the gem we were using before (webpack-rails)
has been unmaintained for quite a while. Thus, migration is the only way forward
and this is what this commit achieves.

All that being said, there are *lots* of rough edges that are pending to be
looked at. I'm going to do this in subsequent commits.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>