frontend: let's fetch the client configuration from a JSON object
We used to render the client configuration into a javascript file, which was
rendered before the main javascript files. This apparently worked just
fine, but this was no longer the case with our move to webpacker.

Instead of doing all that, now the rendered data will be stored as a JSON
object. This object is encoded in base64 just so there are no surprises with
characters and things like that. This is not done for security reasons because
this client configuration should never have sensitive information. Then, before
setting up Vue, this JSON blob will be decoded and parsed so to set the global
configuration object accordingly.

Moreover, as a final touch, note that we are now setting defaults if any of all
this goes wrong, just so the application does not bail completely on an
unexpected situation.

Last but not least, I've recovered the `defer` tag from the javascript tags and
from now on yarn will also print its version on CI.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>